requirements.txt
medpy
nibabel
tensorflow-gpu==2.1.0
tensorflow-addons==0.9.1
h5py==3.1.0
opencv-python
numpy
sklearn
